refactor(repo): move typedoc reflection work into custom-theme, reduce extract-methods to a marker slicer
