+++ .changeset/typedoc-marker-slicer.md
@@ -1,0 +1,2 @@
+---
+---
